Separated complex gemm load store & addressed PR comments
